filterNoiseText, refactor
