+++ app/services/transcription/downloadAndUnzip.ts
@@ -1,0 +1,63 @@
+import { createReadStream, createWriteStream } from 'fs';
+import { promises as fs } from 'node:fs';
+import { Readable } from 'stream';
+import { pipeline } from 'stream/promises';
+import unzip from 'unzip-stream';
+
+export async function downloadAndUnzip(
+  url: string,
+  zipPath: string,
+  extractPath: string,
+  onProgress: (status: { downloaded: number; total: number }) => void,
+) {
+  // 1. Download the zip file
+  const response = await fetch(url);
+  if (!response.ok) {
+    throw new Error(`Failed to download: ${response.statusText}`);
+  }
+  if (!response.body) {
+    throw new Error('Response body is null');
+  }
+
+  const contentLength = response.headers.get('Content-Length');
+  const totalSize = contentLength ? parseInt(contentLength, 10) : 0;
+
+  const fileStream = createWriteStream(zipPath);
+  const reader = response.body.getReader();
+  let progressTimer: NodeJS.Timeout | null = null;
+  const progress: { downloaded: number; total: number } = { downloaded: 0, total: totalSize };
+
+  const downloadStream = new Readable({
+    async read() {
+      const { done, value } = await reader.read();
+      if (done) {
+        this.push(null);
+        return;
+      }
+      progress.downloaded += value.length;
+      if (progressTimer === null) {
+        progressTimer = setInterval(() => {
+          onProgress(progress);
+        }, 1000);
+      }
+      this.push(value);
+    },
+  });
+
+  await pipeline(downloadStream, fileStream);
+  console.log('Download completed:', zipPath);
+  if (progressTimer) {
+    clearInterval(progressTimer);
+    progressTimer = null;
+  }
+  onProgress(progress);
+
+  // 2. Unzip the downloaded file
+  await fs.mkdir(extractPath, { recursive: true });
+  console.log('Extracting to:', extractPath);
+
+  const zipStream = createReadStream(zipPath);
+  await pipeline(zipStream, unzip.Extract({ path: extractPath }));
+
+  console.log('Extraction completed.');
+}
